refactor: Restructure server routes and enhance task management functionality
- Moved various routes to dedicated files for better organization, including attachment, file serving, debug, and health routes.
- Integrated rate limiting middleware for login, registration, and account activation endpoints to improve security.
- Enhanced task retrieval logic to include relationships such as comments, watchers, and collaborators, ensuring comprehensive task data is available.
- Updated task update and reorder functionalities to preserve existing task data during updates, improving user experience.
- Improved comment handling in task details to ensure comments are correctly managed during updates and deletions.
